pyproject.toml
[project]
name = "experimental-wsd"
version = "0.1.0"
description = "Repository for experimenting with Word Sense Disambiguation methods."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bokeh>=3.7.3",
    "datasets>=3.6.0",
    "notebook>=7.4.3",
    "pip>=25.1.1",
    "polars>=1.30.0",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "typer>=0.16.0",
    "wn>0.12.0",
    "torch>=2.7.1",
    "transformers[sentencepiece]>=4.53.2",
    "en-core-web-trf",
    "spacy>=3.8.7",
    "lightning[pytorch-extra]==2.5.3",
    "torchmetrics>=1.7.4",
    "triton>=3.3.1",
    "pyyaml>=6.0.2",
    "omegaconf>=2.3.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]
en-core-web-trf = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0.tar.gz" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
include = ["**/src/**/*.py", "**/tests/**/*.py", "**/training_runs/**/*.py", "**/pyproject.toml"]

[tool.ruff.lint]
# Enable the isort rules.
extend-select = ["I"]

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = ["**/*venv/**/*"]

[dependency-groups]
dev = [
    "pyrefly>=0.24.0",
    "pytest>=8.4.1",
    "pytest-repeat>=0.9.4",
    "pytest-rerunfailures>=15.1",
    "ruff>=0.12.3",
]
